@@ -6,7 +6,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0-alpha06'
+        classpath 'com.android.tools.build:gradle:3.6.0-beta02'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -25,12 +25,17 @@ task clean(type: Delete) {
 }
 
 ext {
-    compileSdkVersion = 28
-    buildToolsVersion = "28.0.3"
+    compileSdkVersion = 29
     minSdkVersion = 14
-    targetSdkVersion = 28
+    targetSdkVersion = 29
 
-    supportLibraryVersion = "28.0.0"
+    androidXVersions = [
+            appCompat  : '1.1.0',
+            annotation : '1.1.0',
+            core       : '1.2.0-beta01',
+    ]
 
-    junitVersion = "4.12"
+    materialLibraryVersion = "1.2.0-alpha01"
+
+    junitVersion = "4.13-beta-2"
 }
