@@ -1,15 +1,14 @@
 package com.badgeview.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
 
 import com.badgeview.BadgeActionMenuItemView;
-import com.badgeview.BadgeView;
+import com.badgeview.sample.databinding.ActivityMainBinding;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -18,17 +17,11 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
 
-        final BadgeView badgeView = findViewById(R.id.test_badge_view);
-        final TextView textView = findViewById(R.id.test_text_view);
+        ActivityMainBinding binding = ActivityMainBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
 
-        textView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                badgeView.toggle();
-            }
-        });
+        binding.testTextView.setOnClickListener(v -> binding.testBadgeView.toggle());
     }
 
     @Override
